Revert "Home page"
