+++ angularFrontend/src/app/feature/admin-area/admin-area.module.ts
@@ -6,10 +6,8 @@ import { CommonModule } from '@angular/common';
 import { RouterModule, Routes } from '@angular/router';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { HostingAnalysisComponent } from './hosting-analysis/hosting-analysis.component';
-import { AuthGuard } from 'src/app/core/auth/auth.guard';
 import { AdminGuard } from 'src/app/core/auth/admin.guard';
-import { HomeComponent } from './home/home.component';
-
+import { HomeAdminComponent } from './home-admin/home-admin.component';
 
 const routes: Routes = [
   {
@@ -25,15 +23,14 @@ const routes: Routes = [
       {
         path: 'home',
         canLoad: [AdminGuard],
-        loadChildren: () => import('./home/home.module').then(m => m.HomeModule),
+        loadChildren: () => import('./home-admin/home-admin.module').then(m => m.HomeAdminModule),
       }
     ]
   }
 ]
 
 @NgModule({
-  declarations: [AdminAreaComponent, HostingAnalysisComponent, HomeComponent],
-
+  declarations: [AdminAreaComponent, HostingAnalysisComponent, HomeAdminComponent],
   imports: [
     CommonModule,
     RouterModule.forChild(routes),
