+++ angularFrontend/src/app/feature/admin-area/home-admin/home-admin.component.html
@@ -1,0 +1,1 @@
+<p>home-admin works!</p>
