+++ angularFrontend/src/app/feature/admin-area/home-admin/home-admin.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-home-admin',
+  templateUrl: './home-admin.component.html',
+  styleUrls: ['./home-admin.component.css']
+})
+export class HomeAdminComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit(): void {
+  }
+
+}
