+++ angularFrontend/src/app/feature/admin-area/home-admin/home-admin.module.ts
@@ -1,0 +1,25 @@
+import { HomeAdminComponent } from './home-admin.component';
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+import { RouterModule, Routes } from '@angular/router';
+import { NgxMaskModule } from 'ngx-mask';
+
+const routes: Routes = [
+  {
+    path: '',
+    component: HomeAdminComponent
+  }
+];
+
+@NgModule({
+  declarations: [],
+  imports: [
+    CommonModule,
+    RouterModule.forChild(routes),
+    FormsModule,
+    ReactiveFormsModule,
+    NgxMaskModule.forRoot(),
+  ]
+})
+export class HomeAdminModule { }
