--- angularFrontend/src/app/feature/admin-area/home/home.component.css
@@ -1,47 +1,0 @@
-.tasks {
-    border: 2px solid #cee891;
-    margin-top: 15px;
-    width: 60%;
-    margin-left: 40px;
-    border-radius: 4px;
-}
-
-.box {
-    background: #dbf0ac;
-    list-style-type: none;
-    margin-bottom: 2%;
-    margin-right: 10%;
-    border-radius: 4px;
-    height: 30px;
-}
-
-.task {
-    text-align: center;
-}
-
-.nav-item {
-    list-style-type: none;
-    margin-top: 15px;
-    width: 70%;
-    margin-left: 20%;
-}
-
-.nav-link {
-    color: black;
-    background: #dbf0ac;
-    text-align: center;
-    border-radius: 4px;
-}
-
-.name {
-    margin-top: 20%;
-    color: black;
-    font-weight: bolder;
-}
-
-.list {
-    list-style-type: none;
-    border: 1px solid #cee891;
-    text-align: center;
-    border-radius: 3px;
-}
