--- angularFrontend/src/app/feature/admin-area/home/home.component.html
@@ -1,68 +1,0 @@
-<div class = "row">
-    <div class = "tasks col-md-8">
-        <div class = "title">
-            <h2>Lembrete!</h2>
-            <ul>
-                <li class = "box">
-                    <div class = "task">Folga na Páscoa</div>
-                </li>
-                <li class = "box">
-                    <div class = "task">Dar banho na Mel</div>
-                </li>
-                <li class = "box">
-                    <div class = "task">Cobrar pagamento de cliente caloteiro</div>
-                </li>
-            </ul>
-        </div>
-    </div>
-    
-    <div class = "nav-bar col-md-4">
-        <ul>
-            <li class="nav-item" >
-                <a class="nav-link">Clientes</a>
-            </li>
-            <li class="nav-item" >
-                <a class="nav-link">Pets</a>
-            </li>
-            <li class="nav-item" >
-                <a class="nav-link" [routerLink]="['/admin-area/hosting-analysis']">Analisar Hospedagem</a>
-            </li>
-            <li class="nav-item" >
-                <a class="nav-link">Status dos Pets</a>
-            </li>
-            <li class="nav-item" >
-                <a class="nav-link">Home</a>
-            </li>
-            <li class="nav-item" >
-                <a class="nav-link">Infraestrutura</a>
-            </li>
-            <li class="nav-item" >
-                <a class="nav-link">Calendário</a>
-            </li>
-        </ul>
-    </div>
-</div>
-
-<div class = "row">
-    <div class = "col-md-4">
-        <ul>
-            <li class = "name list">TO DO</li>
-            <li class = "list">Banho na Mel</li>
-            <li class = "list">Limper o quarto dos passáros grandes</li>
-        </ul>
-    </div>
-    <div class = "col-md-4">
-        <ul>
-            <li class = "name list">DOING</li>
-            <li class = "list">Brincar com Totó</li>
-            <li class = "list">-</li>
-          </ul>
-    </div>
-    <div class = "col-md-4">
-        <ul>
-            <li class = "name list">DONE</li>
-            <li class = "list">Devolver o demoneo pro dono</li>
-            <li class = "list">Chora no banheiro</li>
-          </ul>
-    </div>
-</div>
