--- angularFrontend/src/app/feature/admin-area/home/home.module.ts
@@ -1,20 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { RouterModule, Routes } from '@angular/router';
-import { HomeComponent } from './home.component';
-
-const routes: Routes = [
-  {
-    path: '',
-    component: HomeComponent
-  }
-];
-
-@NgModule({
-  declarations: [],
-  imports: [
-    CommonModule,
-    RouterModule.forChild(routes),
-  ]
-})
-export class HomeModule { }
