+++ angularFrontend/src/app/feature/admin-area/hosting-analysis/hosting-analysis.component.css
@@ -3,8 +3,7 @@
     max-width: 1128px;
     margin-right: auto;
     margin-left: auto;
-    margin-top: 5%;
-    border: 3px solid #cee891;
+    border: 3px solid #BADA5F;
     border-radius: 10px;
     padding: 10px 0 35px 0;
     margin-bottom: 60px;
@@ -21,7 +20,7 @@
 }
 
 .styled-table thead tr {
-    background-color: #bce955;
+    background-color: #BADA5F;
     color: #ffffff;
     text-align: left;
 }
@@ -40,7 +39,7 @@
 }
 
 .styled-table tbody tr:last-of-type {
-    border-bottom: thin solid #bce955;
+    border-bottom: thin solid #BADA5F;
 }
 
 .modal-xl {
