+++ angularFrontend/src/app/feature/auth/login/login.component.ts
@@ -43,7 +43,7 @@ export class LoginComponent implements OnInit {
           localStorage.setItem('token', JSON.stringify(response.token));
           console.log("Deu bom");
           if(response.user.staff){
-            this.router.navigate(['/admin-area/home']);
+            this.router.navigate(['/admin-area/hosting-analysis']);
           }else{
             this.router.navigate(['/user-area/home']);
           }
