+++ angularFrontend/src/app/feature/home/home.component.html
@@ -1,88 +1,2 @@
-<div class = "homeArea">
-    <div class = "header">
-        <div class = "banner">
-            <div class = "banner-img"></div>
-        </div>
-    </div>
-    
-    <div class = "middle">
-        <div class = "row">
-            <div class = "col-md-6">
-                <div class = "x1"></div>
-            </div>
-            <div class = "col-md-6">
-                <div class = "text-box">
-                    <div class = "text">Nós oferecemos serviços de hospedagens. Vai viajar e não sabe com quem deixar seu animalzinho? Deixa conosco! Nossos quartinhos são separados por tamanho e temperamento do bichinho, para não estressá-los e poderem curtir uma boa soneca.</div>
-                </div>
-            </div>
-            <div class = "col-md-6">
-                <div class = "text-box">
-                    <div class = "text">Também garantimos o entretenimento do seu pet. Todo dia atividades solos ou em grupo para promover o bem-estar, saúde e a inteligência do seu bichinho (eles são mais espertos do que a gente imagina).</div>
-                </div>
-            </div>
-            <div class = "col-md-6">
-                <div class = "x2"></div>
-            </div>
-            <div class = "col-md-6">
-                <div class = "x3"></div>
-            </div>
-            <div class = "col-md-6">
-                <div class = "text-box">
-                    <div class = "text">Seu pet precisa de cuidados médicos especiais? Conte para nós na hora de agendar e iremos providenciar atendimento veterinário 24h, já incluso no preço da hospedagem!*</div>
-            </div>
-            </div>
-        </div>
-        <div class = "sm-text">*Atendimento veterinário não discutido durante o agendamento da hospedagem serão cobrados posteriormente.</div>
-    </div>
-    
-    <div class = "footer">
-        <div class = "row">
-            <div class = "col-md-3">
-                <div class="title">Pousada animal</div>
-                <ul class = "list">
-                    <li><a class = "link" href="#">Como Funciona</a></li>
-                    <li><a class = "link" href="#">Onde Estamos</a></li>
-                    <li><a class = "link" href="#">Trabalhe Conosco</a></li>
-                </ul>
-            </div>
-            <div class = "col-md-3">
-                <div class="title">Serviços</div>
-                <ul class = "list">
-                    <li><a class = "link" href="#">Hospedagem</a></li>
-                    <li><a class = "link" href="#">Tosa e banho</a></li>
-                    <li><a class = "link" href="#">Serviço veterinário</a></li>
-                    <li><a class = "link" href="#">Creche</a></li>
-                </ul>
-            </div>
-            <div class = "col-md-3">
-                <div class="title">Central de Ajuda</div>
-                <ul class = "list">
-                    <li><a class = "link" href="#">FAQ</a></li>
-                    <li><a class = "link" href="#">Termos de Uso</a></li>
-                    <li><a class = "link" href="#">Política de Privacidade</a></li>
-                </ul>
-            </div>
-            <div class = "col-md-3">
-                <div class="title">Saiba Mais</div>
-                <ul class = "list">
-                    <li><a class = "link" href="#">Facebook</a></li>
-                    <li><a class = "link" href="#">Instagram</a></li>
-                    <li><a class = "link" href="#">Twitter</a></li>
-                    <li><a class = "link"href="#">Reclame Aqui</a></li>
-                </ul>
-            </div>
-        </div>
-        <div class = "line"></div>
-        <div class = "container-email">
-            <div class = "title-email">Receba as novidades!</div>
-            <div class = "input-email align-items-center">
-                <div class = "email-icon"></div>
-                <div class = "email-text">
-                    <input name="email" type="text" placeholder="E-mail" aria-label="newsletter" class="ng-untouched ng-pristine ng-valid">
-                </div>
-            </div>
-            <button type="submit" class="btn">Enviar</button>
-        </div>
-    </div>
-</div>
-
+<p>home works!</p>
+<h5>TESTE</h5>
