+++ angularFrontend/src/app/feature/nav-bar/nav-bar.component.css
@@ -28,7 +28,7 @@
 
 /* .navbar-img{
   margin: auto;
-}
+};
 
 .navbar-components{
 margin: auto;
@@ -42,6 +42,7 @@ align-content: space-between;
 
 .navbar{
   margin: 0 -15px;
+  margin-bottom: 60px;
   height: 100px;
   padding: 0.4rem 1rem;
   /* background-color: #50c8c6; */
