+++ angularFrontend/src/app/feature/user-area/home-user/home-user.component.css
@@ -3,9 +3,14 @@
 }
 
 .banner{
-    height: 200px;
+    margin-top: -3.9%;
+    height: 20%;
     width: 100%;
-    background-color: #a7cd2c;;
+    background-color: #E1F5C4;
+}
+
+.text{
+  text-align: center;
 }
 
 .icon{
@@ -27,13 +32,13 @@
 
 .edit-bottom{
     background-color: white;
-    border: 2px solid #a7cd2c;;
+    border: 2px solid #A7CD2C;
     border-radius: 5px;
     color: #000000;
 }
 
 .name{
-    margin-top: -4%;
+    margin-top: -5%;
     margin-left: 20%;
     font-size: 20px;
     font-family: 'Gill Sans', 'Gill Sans MT', Calibri, 'Trebuchet MS', sans-serif;
@@ -59,7 +64,7 @@
     margin-top: 20px;
     color: #000000;
     font-size: 15px;
-    border: 2px solid #a7cd2c;
+    border: 2px solid #A7CD2C;
     border-radius: 8px;
 }
 
@@ -71,14 +76,13 @@
     display: flex;
     align-items: center;
     width: 80%;
-    background-color: #d7e99e;;
+    background-color: #E1F5C4;
     border-radius: 5px;
 }
 
 .buttom {
     background-color: white;
-    border: 1px solid #c6ee43;
-
+    border: 1px solid #A7CD2C;
     border-radius: 5px;
     color: black;
     font-size: 12px;
@@ -97,7 +101,7 @@
     align-items: center;
     justify-content: center;
     width: 150px;
-    background-color: #c6ee43;
+    background-color: #A7CD2C;
     border-radius: 5px;
   }
