+++ angularFrontend/src/app/feature/user-area/hosting/hosting.component.css
@@ -10,8 +10,7 @@
   max-width: 1128px;
   margin-right: auto;
   margin-left: auto;
-  margin-top: 20px;
-  border: 3px solid #a7cd2c;
+  border: 3px solid #bada5f;
   border-radius: 10px;
   padding: 50px 30px;
 }
@@ -27,20 +26,20 @@
   height: 48px;
   color: black;
   background-color: #f5fff7;
-  border: 1px solid #a7cd2c;
+  border: 1px solid #bada5f;
   border-radius: 5px;
   outline: none;
 }
 
 .form-select{
-  border: 1px solid #a7cd2c;
+  border: 1px solid #bada5f;
   background-color: #f5fff7;
   width: 100%;
   height: 46px;
 }
 
 .btn-bottom {
-  border: 2px solid #a7cd2c;
+  border: 2px solid #bada5f;
   border-radius: 5px;
   color: #000000;
   margin: 0 auto;
