+++ angularFrontend/src/app/feature/user-area/pet-register/pet-register.component.css
@@ -33,13 +33,13 @@
   width: 100%;
   height: 48px;
   background-color: #f5fff7;
-  border: 1px solid #a7cd2c;
+  border: 1px solid #bada5f;
   border-radius: 4px;
   outline: none;
 }
 
 .register-form input[type='radio']{
-  border: 1px solid #a7cd2c;
+  border: 1px solid #bada5f;
   border-radius: 4px;
   /* background-color:rgb(249, 253, 237); */
   /* background:rgb(249, 253, 237); */
@@ -49,7 +49,7 @@
 }
 
 .register-form textarea {
-  border: 1px solid #a7cd2c;
+  border: 1px solid #bada5f;
   border-radius: 4px;
 }
 
@@ -58,11 +58,10 @@
   max-width: 1128px;
   margin-right: auto;
   margin-left: auto;
-  border: 3px solid #a7cd2c;
+  border: 3px solid #bada5f;
   border-radius: 10px;
   padding: 10px 0 35px 0;
   margin-bottom: 60px;
-  margin-top: 20px;
 }
 
 h2 {
@@ -70,7 +69,7 @@ h2 {
 }
 
 .form-select{
-  border: 1px solid #a7cd2c;
+  border: 1px solid #bada5f;
   background-color: #f5fff7;
   width: 100%;
   height: 46px;
@@ -113,7 +112,7 @@ h2 {
 }
 
 .btn {
-  background-color: #a7cd2c;
+  background-color: #bada5f;
   margin-top: 30px;
   float: right;
 }
